backend/requirements.txt
asgiref==3.5.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.1
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==36.0.2
distlib==0.3.3
Django==4.0.3
django-cors-headers==3.0.2
django-filter==2.4.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djoser==2.0.5
et-xmlfile==1.1.0
filelock==3.2.0
Flask==2.0.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
huggingface-hub==0.0.17
idna==2.8
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.0.1
Markdown==3.3
MarkupSafe==2.0.1
MouseInfo==0.1.3
mysqlclient==2.1.0
numpy==1.21.2
openai==0.14.0
openpyxl==3.0.9
packaging==21.0
pandas==1.4.0
pandas-stubs==1.2.0.47
Pillow==8.3.2
pipenv==2021.11.23
platformdirs==2.4.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.3
PyYAML==5.4.1
regex==2021.9.30
requests==2.22.0
sacremoses==0.0.46
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.31
sqlparse==0.4.2
stripe==3.4.0
torchaudio==0.9.1
torchvision==0.10.1
tqdm==4.62.3
transformers==4.11.2
typing-extensions==3.10.0.2
tzdata==2022.1
urllib3==1.25.11
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.6
waitress==1.4.3
Werkzeug==2.0.3
whitenoise==4.1.4
windows-curses==2.3.0